requirements.txt
astroid==2.4.2
Click==7.0
dominate==2.4.0
Flask==1.1.1
Flask-Bootstrap==3.3.7.1
gunicorn==20.0.4
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.3
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
pylint==2.6.0
PyYaml==5.3.1
six==1.13.0
typed-ast==1.4.1
visitor==0.1.3
Werkzeug==0.16.0
wrapt==1.11.2
